requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
coverage==7.8.0
Django==5.1.7
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.1.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PySocks==1.7.1
pytest==8.3.5
pytest-django==4.11.1
selenium==4.32.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
